+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This release is based on Jujutsu 0.41.
 - Workspaces are now interactive objects. The `@`-chip shown on a workspace's working copy revision has a context menu with two new commands:
   * Rename... - renames the workspace.
   * Forget - removes a workspace from the repository, abandoning its working copy if it's empty.
+- Chips support double-clicking to perform a default action - track/untrack for bookmarks.
 - The Changes line in the right pane shows diff stats for non-empty revisions.
 
 ### Fixed